detailView capsule & Color sum
