+++ checkMate/ViewController/Detail/ContentsDetailViewController.swift
@@ -17,6 +17,7 @@ import gRPC_Core
 
 
 class ContentsDetailViewController: UIViewController, UITableViewDelegate, UITableViewDataSource, UITextViewDelegate, MFMailComposeViewControllerDelegate {
+    //MARK: - Properties
     var send_username : String!
     
     @IBOutlet weak var replyTableView: UITableView!
@@ -33,12 +34,7 @@ class ContentsDetailViewController: UIViewController, UITableViewDelegate, UITab
     
     var replyTableViewController = UITableViewController()
     
-    // 라인 회색컬러
-    let color = UIColor(rgb: 0xE5E5E5)
-
-//    var contensDetailData: Post!
-//    var replySection: Int!
-    
+    let color = UIColor.mainGrey
     
     let db = Firestore.firestore()
     var lef: DatabaseReference!
@@ -47,12 +43,10 @@ class ContentsDetailViewController: UIViewController, UITableViewDelegate, UITab
     var scrapFlag = false
     var check_replyuser : [Bool] = []
     var viewHeight : CGFloat = 0
-    // Add a new document with a generated ID
     var ref: DocumentReference? = nil
-
     var contentsDetailData: Post!
     
-    
+    //MARK: - IBAction
     @IBAction func ScrapButton(_ sender: Any) {
         guard Auth.auth().currentUser != nil else {
             haveUesr()
@@ -91,168 +85,7 @@ class ContentsDetailViewController: UIViewController, UITableViewDelegate, UITab
         
         
     }
-
     
-    override func viewWillAppear(_ animated: Bool) {
-        super.viewWillAppear(animated)
-      
-        DataLoad()
-        replyTableView.reloadData()
-        habitDataLoad()
-        // 게시글 작성자만 삭제 버튼 보이게 하기
-        if Auth.auth().currentUser == nil {
-            contentDeleteButton.isHidden = true
-//            deleteButton.isHidden = true
-        } else {
-            if Auth.auth().currentUser?.uid != contentsDetailData.uid {
-//              deleteButton.isHidden = true
-                contentDeleteButton.isHidden = true
-                
-            }
-        }
-        
-       
-    }
-    
-    override func viewWillDisappear(_ animated: Bool) {
-            super.viewWillDisappear(animated)
-            
-            NotificationCenter.default.removeObserver(self)
-    }
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        
-        //옵저버
-        NotificationCenter.default.addObserver(self, selector: #selector(keyboardWillShow), name: UIResponder.keyboardWillShowNotification, object: nil)
-        NotificationCenter.default.addObserver(self, selector: #selector(keyboardWillHide), name: UIResponder.keyboardWillHideNotification, object: nil)
-        
-        viewHeight = self.view.frame.size.height
-        // 키보드 코드
-         
-       
-//        self.replyTableView.frame.inset(by: UIEdgeInsets(top: 0, left: 15, bottom: 0, right: 15))
-
-//        writerPatternButton.backgroundColor = .white
-//        writerPatternButton.tintColor = .white
-//        writerPatternButton.layer.borderWidth = 1
-//        writerPatternButton.layer.borderColor = color.cgColor
-//        writerPatternButton.layer.cornerRadius = 15
-        replyOkButton.layer.borderColor = color.cgColor
-        replyOkButton.layer.borderWidth = 1
-        replyOkButton.layer.cornerRadius = 10
-        replyOkButton.tintColor = .white
-        replyTextField.layer.borderWidth = 1
-        replyTextField.layer.cornerRadius = 10
-        replyTextField.layer.borderColor = color.cgColor
-        
-        currentData = contentsDetailData
-        userHabitCheck.removeAll()
-        
-
-        
-        scrapDataLoad{ (result) in
-            print("*클로저 실행\(result)")
-            self.scrapFlag = result
-            if self.scrapFlag{
-                print("*별색 채우기")
-                self.navigationItem.rightBarButtonItem?.image = UIImage(systemName: "star.fill")
-            }else{
-                print("*별색 못채우기")
-                self.navigationItem.rightBarButtonItem?.image = UIImage(systemName: "star")
-            }
-        }
-        
-//        lef = Database.database().reference(withPath: "servey")
-        
-        let userLabel = self.view.viewWithTag(1) as! UILabel
-        let tittleLabel = self.view.viewWithTag(2) as! UILabel
-        let contentsText = self.view.viewWithTag(3) as! UITextView
-        let dateLabel = self.view.viewWithTag(4) as! UILabel
-        
-        replyTableViewController.tableView.delegate = self
-        replyTableViewController.tableView.dataSource = self
-        replyTableView.delegate = self
-        replyTableView.dataSource = self
-        
-
-        userLabel.text = contentsDetailData.author
-        send_username =  userLabel.text
-        userLabel.textAlignment = .right
-        userLabel.sizeToFit()
-        tittleLabel.text = contentsDetailData.title
-        tittleLabel.sizeToFit()
-        contentsText.text = contentsDetailData.contents
-        contentsText.sizeToFit()
-        dateLabel.text = contentsDetailData.date
-        dateLabel.sizeToFit()
-        // Do any additional setup after loading the view.
-        
-        // 셀 크기 자동조절
-        replyTableView.estimatedRowHeight = 60
-        replyTableView.rowHeight = UITableView.automaticDimension
-    }
-    
-    func numberOfSections(in tableView: UITableView) -> Int {
-        return replyList.count
-    }
-    
-    func tableView(_ tableView: UITableView, heightForFooterInSection section: Int) -> CGFloat {
-        return 15
-    }
-    
-    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        return 1
-    }
-    
-    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        let cell = replyTableView.dequeueReusableCell(withIdentifier: "replyCell", for: indexPath)
-        
-        
-        //  둥근 테두리 만들기
-        cell.backgroundColor = UIColor.white
-        cell.layer.borderColor = color.cgColor
-        cell.layer.borderWidth = 1
-        cell.layer.cornerRadius = 8
-        cell.clipsToBounds = true
-//        cell.frame.inset(by: UIEdgeInsets(top: 0, left: 15, bottom: 0, right: 15))
-        
-        let userLabel = cell.viewWithTag(1) as! UILabel
-        let contentsText = cell.viewWithTag(2) as! UITextView
-        let dateLabel = cell.viewWithTag(3) as! UILabel
-        let replyDeleteLabel = cell.viewWithTag(4) as! UIButton
-        
-
-        if replyList.count > 0{
-            // 텍스트 크기 자동조절
-            userLabel.text = self.replyList[indexPath.section].author
-            userLabel.sizeToFit()
-            contentsText.text = self.replyList[indexPath.section].contents
-            contentsText.sizeToFit()
-            dateLabel.text = self.replyList[indexPath.section].date
-            dateLabel.sizeToFit()
-            
-            // 삭제 버튼 확이
-            replyDeleteLabel.isHidden = true
-            if Auth.auth().currentUser == nil || Auth.auth().currentUser?.uid != self.replyList[indexPath.section].uid   {
-                replyDeleteLabel.isHidden = true
-            }
-            print("$$$\(self.replyList[indexPath.section].uid): \(Auth.auth().currentUser?.uid)")
-            if self.check_replyuser[indexPath.section] {
-                replyDeleteLabel.isHidden = false
-            }
-//            if self.replyList[indexPath.section].uid != Auth.auth().currentUser?.uid {
-//                replyDeleteLabel.isHidden = true
-//            }
-//            else {}
-            
-        }
-        else {}
-        
-        return cell
-    }
-
-
     @IBAction func goTotresh(_ sender: Any) {
         print("삭제")
         let db = Firestore.firestore()
@@ -321,16 +154,6 @@ class ContentsDetailViewController: UIViewController, UITableViewDelegate, UITab
             DataLoad()
             self.replyTableView.reloadData()
             
-            // 화면 리로드
-//            viewWillAppear(true)
-
-//            ref = db.collection("Post").document().collection("Post").document().setData() { err in
-//                if let err = err {
-//                    print("Error writing document: \(err)")
-//                } else {
-//                    print("Document successfully written!")
-//                }
-//            }
             
         }
         if AppDelegate.user == nil {
@@ -344,6 +167,153 @@ class ContentsDetailViewController: UIViewController, UITableViewDelegate, UITab
             self.replyTableView.reloadData()
         }
     }
+    
+    /// 댓글 삭제
+    @IBAction func replyDelete(_ sender: UIButton) {
+        // 버튼의 위치 찾기
+        let contentView = sender.superview
+        let cell = contentView?.superview as! UITableViewCell
+        let indexPath = replyTableView.indexPath(for: cell)
+        
+        if replyList.count > 0 {
+            // 댓글의 문서 id 삭제
+            db.collection("Post").document(contentsDetailData.pid).collection("Comment").document(replyList[indexPath!.section].docid).delete() { err in
+                if let err = err {
+                    print("Error removing document: \(err)")
+                } else {
+                    print("Document successfully removed!")
+                }
+            }
+        }
+
+        
+        // 화면 재구성
+        replyList.removeAll()
+        DataLoad()
+        self.replyTableView.reloadData()
+        
+    }
+    
+    /// 게시글 신고하기
+    @IBAction func reportButton(_ sender: Any) {
+        
+        // 이메일 사용가능한지 체크하는 if문
+        if MFMailComposeViewController.canSendMail() {
+            
+            let compseVC = MFMailComposeViewController()
+            compseVC.mailComposeDelegate = self
+            
+            compseVC.setToRecipients(["deeeeeep0122@gmail.com"])
+            compseVC.setSubject("[메이트리스트]게시글 신고")
+            compseVC.setMessageBody("\(contentsDetailData.pid)", isHTML: false)
+            self.present(compseVC, animated: true, completion: nil)
+            
+        }
+        else {
+            self.showSendMailErrorAlert()
+        }
+    }
+    
+    @IBAction func patternButtonClick(_ sender: Any){
+        self.view.endEditing(true)
+    }
+    
+    
+    //MARK: - LifeCycle
+    override func viewWillAppear(_ animated: Bool) {
+        super.viewWillAppear(animated)
+      
+        DataLoad()
+        replyTableView.reloadData()
+        habitDataLoad()
+        // 게시글 작성자만 삭제 버튼 보이게 하기
+        if Auth.auth().currentUser == nil {
+            contentDeleteButton.isHidden = true
+        } else {
+            if Auth.auth().currentUser?.uid != contentsDetailData.uid {
+                contentDeleteButton.isHidden = true
+                
+            }
+        }
+        
+       
+    }
+    
+    override func viewWillDisappear(_ animated: Bool) {
+            super.viewWillDisappear(animated)
+            
+            NotificationCenter.default.removeObserver(self)
+    }
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        
+        //옵저버
+        NotificationCenter.default.addObserver(self, selector: #selector(keyboardWillShow), name: UIResponder.keyboardWillShowNotification, object: nil)
+        NotificationCenter.default.addObserver(self, selector: #selector(keyboardWillHide), name: UIResponder.keyboardWillHideNotification, object: nil)
+        
+        setData()
+        setView()
+ 
+    }
+    //MARK: - Set Function
+    func setData() {
+          currentData = contentsDetailData
+          userHabitCheck.removeAll()
+          
+          scrapDataLoad{ (result) in
+              print("*클로저 실행\(result)")
+              self.scrapFlag = result
+              if self.scrapFlag{
+                  print("*별색 채우기")
+                  self.navigationItem.rightBarButtonItem?.image = UIImage(systemName: "star.fill")
+              }else{
+                  print("*별색 못채우기")
+                  self.navigationItem.rightBarButtonItem?.image = UIImage(systemName: "star")
+              }
+          }
+          
+    }
+    
+    func setView() {
+        viewHeight = self.view.frame.size.height
+       
+        replyOkButton.layer.borderColor = color.cgColor
+        replyOkButton.layer.borderWidth = 1
+        replyOkButton.layer.cornerRadius = 10
+        replyOkButton.tintColor = .white
+        replyTextField.layer.borderWidth = 1
+        replyTextField.layer.cornerRadius = 10
+        replyTextField.layer.borderColor = color.cgColor
+        
+        let userLabel = self.view.viewWithTag(1) as! UILabel
+        let tittleLabel = self.view.viewWithTag(2) as! UILabel
+        let contentsText = self.view.viewWithTag(3) as! UITextView
+        let dateLabel = self.view.viewWithTag(4) as! UILabel
+        
+        replyTableViewController.tableView.delegate = self
+        replyTableViewController.tableView.dataSource = self
+        replyTableView.delegate = self
+        replyTableView.dataSource = self
+        
+
+        userLabel.text = contentsDetailData.author
+        send_username =  userLabel.text
+        userLabel.textAlignment = .right
+        userLabel.sizeToFit()
+        tittleLabel.text = contentsDetailData.title
+        tittleLabel.sizeToFit()
+        contentsText.text = contentsDetailData.contents
+        contentsText.sizeToFit()
+        dateLabel.text = contentsDetailData.date
+        dateLabel.sizeToFit()
+      
+        // 셀 크기 자동조절
+        replyTableView.estimatedRowHeight = 60
+        replyTableView.rowHeight = UITableView.automaticDimension
+    }
+    
+    //MARK: - Custom Function
     
     func DataLoad()  {
             //데이터 불러오기
@@ -375,6 +345,7 @@ class ContentsDetailViewController: UIViewController, UITableViewDelegate, UITab
             self.replyTableView.reloadData()
         }
     }
+    
     func replyUserCheck() {
         self.check_replyuser.removeAll()
         for item in replyList {
@@ -383,8 +354,6 @@ class ContentsDetailViewController: UIViewController, UITableViewDelegate, UITab
             }else{self.check_replyuser.append(true)}
         }
     }
-    
-    
     
     func scrapDataLoad(_ escapingHandler : @escaping (Bool) -> ()) {
             //데이터 불러오기
@@ -439,63 +408,15 @@ class ContentsDetailViewController: UIViewController, UITableViewDelegate, UITab
         
     }
     
-    // 테이블뷰 당겼을때 데이터 새로 불러오기
+    /// 테이블뷰 당겼을때 데이터 새로 불러오기
     func scrollViewDidEndDragging(_ scrollView: UIScrollView, willDecelerate decelerate: Bool) {
         self.replyTableView.viewWithTag(4)?.isHidden = true
 //        replyList.removeAll()
         DataLoad()
         
     }
-    // 댓글 삭제
-    @IBAction func replyDelete(_ sender: UIButton) {
-        // 버튼의 위치 찾기
-        let contentView = sender.superview
-        let cell = contentView?.superview as! UITableViewCell
-        let indexPath = replyTableView.indexPath(for: cell)
-        
-        if replyList.count > 0 {
-            // 댓글의 문서 id 삭제
-            db.collection("Post").document(contentsDetailData.pid).collection("Comment").document(replyList[indexPath!.section].docid).delete() { err in
-                if let err = err {
-                    print("Error removing document: \(err)")
-                } else {
-                    print("Document successfully removed!")
-                }
-            }
-        }
-
-        
-        // 화면 재구성
-        replyList.removeAll()
-        DataLoad()
-        self.replyTableView.reloadData()
-        
-    }
-    
-    // 게시글 신고하기
-    @IBAction func reportButton(_ sender: Any) {
-        
-        // 이메일 사용가능한지 체크하는 if문
-        if MFMailComposeViewController.canSendMail() {
-            
-            let compseVC = MFMailComposeViewController()
-            compseVC.mailComposeDelegate = self
-            
-            compseVC.setToRecipients(["deeeeeep0122@gmail.com"])
-            compseVC.setSubject("[메이트리스트]게시글 신고")
-            compseVC.setMessageBody("\(contentsDetailData.pid)", isHTML: false)
-            self.present(compseVC, animated: true, completion: nil)
-            
-        }
-        else {
-            self.showSendMailErrorAlert()
-        }
-    }
-    @IBAction func patternButtonClick(_ sender: Any){
-        self.view.endEditing(true)
-    }
-    
-    // 메일 에러 메시지
+  
+    /// 메일 에러 메시지
     func showSendMailErrorAlert() {
             let sendMailErrorAlert = UIAlertController(title: "메일을 전송 실패", message: "아이폰 이메일 설정을 확인하고 다시 시도해주세요.", preferredStyle: .alert)
             let confirmAction = UIAlertAction(title: "확인", style: .default) {
@@ -506,13 +427,10 @@ class ContentsDetailViewController: UIViewController, UITableViewDelegate, UITab
             self.present(sendMailErrorAlert, animated: true, completion: nil)
         }
     
-    // 메일 보낸 뒤 메일 창 닫기
+    /// 메일 보낸 뒤 메일 창 닫기
     func mailComposeController(_ controller: MFMailComposeViewController, didFinishWith result: MFMailComposeResult, error: Error?) {
             controller.dismiss(animated: true, completion: nil)
         }
-    
-    
-    
     
     func habitDataLoad()  {
             //데이터 불러오기
@@ -542,6 +460,69 @@ class ContentsDetailViewController: UIViewController, UITableViewDelegate, UITab
             }
         }
     }
+    
+    //MARK: - Reply TableView
+    func numberOfSections(in tableView: UITableView) -> Int {
+        return replyList.count
+    }
+    
+    func tableView(_ tableView: UITableView, heightForFooterInSection section: Int) -> CGFloat {
+        return 15
+    }
+    
+    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
+        return 1
+    }
+    
+    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
+        let cell = replyTableView.dequeueReusableCell(withIdentifier: "replyCell", for: indexPath)
+        
+        
+        //  둥근 테두리 만들기
+        cell.backgroundColor = UIColor.white
+        cell.layer.borderColor = color.cgColor
+        cell.layer.borderWidth = 1
+        cell.layer.cornerRadius = 8
+        cell.clipsToBounds = true
+//        cell.frame.inset(by: UIEdgeInsets(top: 0, left: 15, bottom: 0, right: 15))
+        
+        let userLabel = cell.viewWithTag(1) as! UILabel
+        let contentsText = cell.viewWithTag(2) as! UITextView
+        let dateLabel = cell.viewWithTag(3) as! UILabel
+        let replyDeleteLabel = cell.viewWithTag(4) as! UIButton
+        
+
+        if replyList.count > 0{
+            // 텍스트 크기 자동조절
+            userLabel.text = self.replyList[indexPath.section].author
+            userLabel.sizeToFit()
+            contentsText.text = self.replyList[indexPath.section].contents
+            contentsText.sizeToFit()
+            dateLabel.text = self.replyList[indexPath.section].date
+            dateLabel.sizeToFit()
+            
+            // 삭제 버튼 확이
+            replyDeleteLabel.isHidden = true
+            if Auth.auth().currentUser == nil || Auth.auth().currentUser?.uid != self.replyList[indexPath.section].uid   {
+                replyDeleteLabel.isHidden = true
+            }
+            print("$$$\(self.replyList[indexPath.section].uid): \(Auth.auth().currentUser?.uid)")
+            if self.check_replyuser[indexPath.section] {
+                replyDeleteLabel.isHidden = false
+            }
+//            if self.replyList[indexPath.section].uid != Auth.auth().currentUser?.uid {
+//                replyDeleteLabel.isHidden = true
+//            }
+//            else {}
+            
+        }
+        else {}
+        
+        return cell
+    }
+
+    //MARK: - 화면 전환
+    
     override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
         
         
@@ -555,13 +536,15 @@ class ContentsDetailViewController: UIViewController, UITableViewDelegate, UITab
         else {}
     }
     
-    // MARK: textview 높이 자동조절
-        func textViewDidChange(_ textView: UITextView) {
+    // MARK: - 키보드 세팅
+    /// textview 높이 자동조절
+    func textViewDidChange(_ textView: UITextView) {
             
             let size = CGSize(width: view.frame.width, height: .infinity)
             let estimatedSize = textView.sizeThatFits(size)
 
-        }
+    }
+    
     @objc func keyboardWillShow(notification: NSNotification) {
 
         //화면-키보드 값 계산
